refactor: moved Navbar.tsx to Navbar folder
