+++ packages/grant-explorer/src/features/common/Navbar/index.ts
@@ -1,0 +1,2 @@
+export { type NavbarProps } from "./Navbar";
+export { default } from "./Navbar";
